requirements.txt
backcall==0.2.0
biopython==1.81
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
fonttools==4.43.0
hurry.filesize==0.9
ipykernel==6.25.2
ipython==7.34.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==5.4.0
kaleido==0.2.1
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.17.0
prettytable==3.9.0
prompt-toolkit==3.0.39
psutil==5.9.5
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
spyder-kernels==2.3.3
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.11.2
tzdata==2023.3
wcwidth==0.2.8
